OrderController implemeted and code reformat done

+++ src/main/java/com/example/dreamshops/dto/ProductDTO.java
@@ -1,19 +1,12 @@
 package com.example.dreamshops.dto;
 
-import com.example.dreamshops.model.Category;
-import com.example.dreamshops.model.Image;
 import com.example.dreamshops.model.Product;
-import jakarta.persistence.CascadeType;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToMany;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.math.BigDecimal;
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Data
 @NoArgsConstructor
@@ -24,7 +17,7 @@ public class ProductDTO {
     private String name;
     private String description;
     private BigDecimal price;
-    private int inventory;
+    private Long inventory;
 
     private CategoryDTO category;
     private List<ImageDTO> images;
